+++ test/Main.hs
@@ -343,7 +343,7 @@ autocommitTest td = testCase "Autocommit mode works"
   $ do
     let sint = Identity (1 :: Int32)
     runQuery_ $ rawSQL "INSERT INTO test1_ (a) VALUES ($1)" sint
-    withNewConnection $ do
+    withNewSession $ do
       n <- runQuery $ rawSQL "SELECT a FROM test1_ WHERE a = $1" sint
       assertEqualEq "Other connection sees autocommited data" 1 n
     runQuery_ $ rawSQL "DELETE FROM test1_ WHERE a = $1" sint
@@ -452,7 +452,7 @@ savepointTest td = testCase "Savepoint support works"
 notifyTest :: TestData -> Test
 notifyTest td = testCase "Notifications work" . runTestEnv td defaultTransactionSettings . unsafeWithoutTransaction $ do
   listen chan
-  forkNewConn $ notify chan payload
+  forkNewSession $ notify chan payload
   mnt1 <- getNotification 250000
   liftBase $ assertBool "Notification received" (isJust mnt1)
   Just nt1 <- pure mnt1
@@ -460,25 +460,25 @@ notifyTest td = testCase "Notifications work" . runTestEnv td defaultTransaction
   assertEqualEq "Payloads are equal" payload (ntPayload nt1)
 
   unlisten chan
-  forkNewConn $ notify chan payload
+  forkNewSession $ notify chan payload
   mnt2 <- getNotification 250000
   assertEqualEq "No notification received after unlisten" Nothing mnt2
 
   listen chan
   unlistenAll
-  forkNewConn $ notify chan payload
+  forkNewSession $ notify chan payload
   mnt3 <- getNotification 250000
   assertEqualEq "No notification received after unlistenAll" Nothing mnt3
   where
     chan = "test_channel"
     payload = "test_payload"
-    forkNewConn action = do
+    forkNewSession action = do
       sem <- newEmptyMVar
-      void . fork . withNewConnection $ do
-        -- withNewConnection needs access to the connection state to get current
+      void . fork . withNewSession $ do
+        -- withNewSession needs access to the connection state to get current
         -- ConnectionAcquisitionMode, but getNotification called immediately
         -- after takes ownership of the connection state for its duration, so if
-        -- CPU gets to it first, withNewConnection will block and notification
+        -- CPU gets to it first, withNewSession will block and notification
         -- will never be sent.
         putMVar sem ()
         action
@@ -496,7 +496,7 @@ transactionTest td lvl =
     $ do
       let sint = Identity (5 :: Int32)
       runQuery_ $ rawSQL "INSERT INTO test1_ (a) VALUES ($1)" sint
-      withNewConnection $ do
+      withNewSession $ do
         n <- runQuery $ rawSQL "SELECT a FROM test1_ WHERE a = $1" sint
         assertEqualEq "Other connection doesn't see uncommited data" 0 n
       rollback
@@ -675,7 +675,7 @@ threadBindingTest td =
       _ <- fork $ do
         sameSessionQuery <- try $ runSQL_ "SELECT 1"
         sameSessionModeChange <- try unsafeAcquireOnDemandConnection
-        newSession <- try . withNewConnection $ runSQL_ "SELECT 1"
+        newSession <- try . withNewSession $ runSQL_ "SELECT 1"
         putMVar result (sameSessionQuery, sameSessionModeChange, newSession)
       (sameSessionQuery, sameSessionModeChange, newSession) <- takeMVar result
       liftBase $ do
@@ -683,7 +683,7 @@ threadBindingTest td =
         assertThreadMismatch "unsafeAcquireOnDemandConnection" sameSessionModeChange
         case newSession of
           Left (e :: SomeException) ->
-            assertFailure $ "withNewConnection failed in another thread: " ++ show e
+            assertFailure $ "withNewSession failed in another thread: " ++ show e
           Right () -> pure ()
   where
     assertThreadMismatch :: String -> Either SomeException () -> Assertion